feat(cubeproxy): add path-based sandbox routing and shared backend resolution
- Introduce`/sandbox/<id>/<port>/...`path-based routing in CubeProxy
- Extract backend resolution and health check into reusable`sandbox_backend.lua`
- Update nginx.conf with new location for both HTTP and HTTPS listeners
- Document the new routing mode in architecture overview and user guides (en/zh)
