+++ CubeProxy/nginx.conf
@@ -90,6 +90,49 @@ http {
         listen 8081 reuseport;
         server_name _;
         set $cube_proxy_host_ip "";
+        # Populated by lua/path_rewrite_phase.lua for the /sandbox/<id>/<port>/ location.
+        # Declared at the server level so proxy_redirect / proxy_cookie_path can reference them.
+        set $ins_id "";
+        set $container_port "";
+
+        # Path-based routing: /sandbox/<sandbox-id>/<container-port>/<rest>
+        # Lets clients reach a sandbox via a plain IP+port without wildcard DNS or TLS.
+        location ^~ /sandbox/ {
+            include /usr/local/openresty/nginx/conf/global/global.conf;
+            set $cube_retcode "310200";
+            set $garyscale_test "none";
+            set $backend_ip "";
+            set $backend_port "";
+            set $access_time "";
+            set $cache_free_space "";
+            set $host_proxy_port 8081;
+
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection "upgrade";
+
+            proxy_send_timeout 7206s;
+            proxy_read_timeout 7206s;
+            proxy_connect_timeout 3s;
+            proxy_set_header Host $http_host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Prefix /sandbox/$ins_id/$container_port;
+
+            # Rewrite root-relative Location headers and cookie Path values back
+            # under the /sandbox/<id>/<port>/ prefix so the client stays scoped.
+            proxy_redirect    ~^/(.*)$  /sandbox/$ins_id/$container_port/$1;
+            proxy_cookie_path /         /sandbox/$ins_id/$container_port/;
+
+            rewrite_by_lua_file lua/path_rewrite_phase.lua;
+
+            proxy_pass http://backend;
+
+            header_filter_by_lua_file lua/header_filter_phase.lua;
+
+            log_by_lua_file lua/log_phase.lua;
+        }
+
         location / {
             include /usr/local/openresty/nginx/conf/global/global.conf;
             set $cube_retcode "310200";
@@ -125,8 +168,51 @@ http {
         listen 8080 ssl reuseport;
         server_name _;
         set $cube_proxy_host_ip "";
+        # Populated by lua/path_rewrite_phase.lua for the /sandbox/<id>/<port>/ location.
+        # Declared at the server level so proxy_redirect / proxy_cookie_path can reference them.
+        set $ins_id "";
+        set $container_port "";
         ssl_certificate     /usr/local/openresty/nginx/certs/cube.app+3.pem;
         ssl_certificate_key /usr/local/openresty/nginx/certs/cube.app+3-key.pem;
+
+        # Path-based routing: /sandbox/<sandbox-id>/<container-port>/<rest>
+        # Lets clients reach a sandbox via a plain IP+port without wildcard DNS or TLS.
+        location ^~ /sandbox/ {
+            include /usr/local/openresty/nginx/conf/global/global.conf;
+            set $cube_retcode "310200";
+            set $garyscale_test "none";
+            set $backend_ip "";
+            set $backend_port "";
+            set $access_time "";
+            set $cache_free_space "";
+            set $host_proxy_port 8080;
+
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection "upgrade";
+
+            proxy_send_timeout 7206s;
+            proxy_read_timeout 7206s;
+            proxy_connect_timeout 3s;
+            proxy_set_header Host $http_host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Prefix /sandbox/$ins_id/$container_port;
+
+            # Rewrite root-relative Location headers and cookie Path values back
+            # under the /sandbox/<id>/<port>/ prefix so the client stays scoped.
+            proxy_redirect    ~^/(.*)$  /sandbox/$ins_id/$container_port/$1;
+            proxy_cookie_path /         /sandbox/$ins_id/$container_port/;
+
+            rewrite_by_lua_file lua/path_rewrite_phase.lua;
+
+            proxy_pass http://backend;
+
+            header_filter_by_lua_file lua/header_filter_phase.lua;
+
+            log_by_lua_file lua/log_phase.lua;
+        }
+
         location / {
             include /usr/local/openresty/nginx/conf/global/global.conf;
             set $cube_retcode "310200";
